Added more tests, added sinon in dev
