+++ app/Utils/testUtils.js
@@ -1,0 +1,28 @@
+process.env.NODE_ENV = 'test';
+
+const User = require('../Models/user');
+const appConstants = require('../Constant/constants');
+const bcrypt = require('bcryptjs');
+const chai = require('chai');
+const chaiHttp = require('chai-http');
+chai.use(chaiHttp);
+
+module.exports = {
+    setupTestUsers: () => {
+        User.remove({}, (err) => { });
+        let hashPwd = bcrypt.hashSync(appConstants.TestUser.password, Number(process.env.SALT));
+        let adminUser = new User({
+            username: appConstants.TestUser.username,
+            password: hashPwd,
+            admin: appConstants.TestUser.admin
+        });
+
+        User.create(adminUser, (err, user) => {
+            if (err) {
+                console.log('Test user setup failed');
+            } else {
+                console.log(`Test user setup successful, Id = ${user.id}`);
+            }
+        });
+    }
+};
