+++ app/test/authentication.js
@@ -1,10 +1,9 @@
 process.env.NODE_ENV = 'test';
 
 const mongoose = require('mongoose');
-const User = require('../Models/user');
 const server = require('../../server');
 const appConstants = require('../Constant/constants');
-const bcrypt = require('bcryptjs');
+const testUtils = require('../Utils/testUtils');
 const chai = require('chai');
 const { expect } = require('chai')
 const chaiHttp = require('chai-http');
@@ -16,25 +15,9 @@ chai.use(chaiHttp);
 
 //Add an Admin users for testing. 
 describe('Authentication', () => {
-    beforeEach((done) => { //Before each test we empty the database
-        User.remove({}, (err) => {
-            done();
-        });
-
-        let hashPwd = bcrypt.hashSync(appConstants.TestUser.password, Number(process.env.SALT));
-        let adminUser = new User({
-            username: appConstants.TestUser.username,
-            password: hashPwd,
-            admin: appConstants.TestUser.admin
-        });
-
-        User.create(adminUser, (err, user) => {
-            if (err) {
-                console.log('Test user setup failed');
-            } else {
-                console.log(`Test user setup successful, Id = ${user.id}`);
-            }
-        });
+    before((done) => { //Before each test we empty the database
+        testUtils.setupTestUsers();
+        done();
     });
 
     describe('/api/login user with incorrect req body', () => {
