+++ package.json
@@ -19,7 +19,8 @@
   "devDependencies": {
     "chai": "^3.5.0",
     "chai-http": "^2.0.1",
-    "mocha": "^2.4.5"
+    "mocha": "^2.4.5",
+    "sinon": "^4.5.0"
   },
   "scripts": {
     "start": "SET NODE_ENV=dev && node server.js",
